+++ abci/did/did.go
@@ -45,12 +45,15 @@ type DIDApplicationInterface struct {
 
 func NewDIDApplicationInterface() *DIDApplicationInterface {
 	logger := logrus.WithFields(logrus.Fields{"module": "abci-app"})
-	var dbDir = getEnv("DB_NAME", "DID")
+
+	var dbType = getEnv("ABCI_DB_TYPE", "cleveldb")
+	var dbDir = getEnv("ABCI_DB_DIR_PATH", "./DID")
+
 	if err := cmn.EnsureDir(dbDir, 0700); err != nil {
 		panic(fmt.Errorf("Could not create DB directory: %v", err.Error()))
 	}
 	name := "didDB"
-	db := dbm.NewDB(name, "cleveldb", dbDir)
+	db := dbm.NewDB(name, dbm.DBBackendType(dbType), dbDir)
 	tree := iavl.NewMutableTree(db, 0)
 	tree.Load()